+++ doc/source/changelog.rst
@@ -3,6 +3,15 @@
 Changelog
 =========
 
+1.4.6 (13 Nov 2020)
+-------------------
+
+Bug fixes
+~~~~~~~~~
+
+- Added `shapely.geometry.MultiPolygon` as an accepted data type in :func:`geowombat.core.prepare_points`.
+- Added missing `GeoDataFrame` columns in :func:`geowombat.core.polygons_to_points`.
+
 1.4.5 (9 Nov 2020)
 ------------------